(WIP) move redefinition of java.base module into agent
