+++ org.jacoco.agent.rt/src/org/jacoco/agent/rt/internal/PreMain.java
@@ -12,6 +12,9 @@
 package org.jacoco.agent.rt.internal;
 
 import java.lang.instrument.Instrumentation;
+import java.util.Collections;
+import java.util.Map;
+import java.util.Set;
 
 import org.jacoco.core.runtime.AgentOptions;
 import org.jacoco.core.runtime.ClassInjectionRuntime;
@@ -54,12 +57,48 @@ public final class PreMain {
 	private static IRuntime createRuntime(final Instrumentation inst)
 			throws Exception {
 
-		final IRuntime runtime = ClassInjectionRuntime.create(inst);
-		if (runtime != null) {
-			return runtime;
+		if (redefineJavaBaseModule(inst)) {
+			return new ClassInjectionRuntime(Object.class, "$JaCoCo");
 		}
 
 		return ModifiedSystemClassRuntime.createFor(inst, "java/lang/UnknownError");
+	}
+
+	private static boolean redefineJavaBaseModule(
+			final Instrumentation instrumentation) throws Exception {
+		try {
+			Class.forName("java.lang.Module");
+		} catch (final ClassNotFoundException e) {
+			return false;
+		}
+
+		Instrumentation.class.getMethod("redefineModule", //
+				Class.forName("java.lang.Module"), //
+				Set.class, //
+				Map.class, //
+				Map.class, //
+				Set.class, //
+				Map.class //
+		).invoke(instrumentation, // instance
+				getModule(Object.class), // module
+				Collections.emptySet(), // extraReads
+				Collections.emptyMap(), // extraExports
+				Collections.singletonMap("java.lang",
+						Collections.singleton(
+								getModule(ClassInjectionRuntime.class))), // extraOpens
+				Collections.emptySet(), // extraUses
+				Collections.emptyMap() // extraProvides
+		);
+		return true;
+	}
+
+	/**
+	 * @return {@code cls.getModule()}
+	 */
+	private static Object getModule(final Class<?> cls) throws Exception {
+		return Class.class //
+				.getMethod("getModule") //
+				.invoke(cls);
 	}
 
 }
